WIP: Show all open tilesets in the Tilesets view
Now the tileset view shows all loaded tilesets, except for embedded
tilesets that are not part of the current map.

Tiled will automatically manage which tilesets are used by a map or not.
Some new action will need to be added for the purpose of removing a
certain tileset from a map.

The tilesets are shown ordered by name. Manual reordering is no longer
possible.

FIXME: Somehow when you open an external tileset, it appears a second
time in the Tilesets view. Is another TilesetDocument created?